Load algorithms from json files
