+++ ai/src/algorithms/dql/dql.json
@@ -1,0 +1,5 @@
+{
+    "algorithm_id": "dql",
+    "name": "Deep Q-Learning",
+    "docs_link": "https://docs.spiceai.org/deep-learning-ai/dql/"
+}
