+++ ai/src/algorithms/vpg/vpg.json
@@ -1,0 +1,5 @@
+{
+    "algorithm_id": "vpg",
+    "name": "Vanilla Policy Gradient",
+    "docs_link": "https://docs.spiceai.org/deep-learning-ai/vpg/"
+}
