+++ pkg/http/server.go
@@ -534,6 +534,19 @@ func apiPostImportHandler(ctx *fasthttp.RequestCtx) {
 	ctx.Response.SetStatusCode(200)
 }
 
+func (server *server) apiGetAlgorithmsHandler(ctx *fasthttp.RequestCtx) {
+	data, err := json.Marshal(aiengine.Algorithms())
+	if err != nil {
+		ctx.Response.SetStatusCode(http.StatusInternalServerError)
+		ctx.Response.SetBodyString(err.Error())
+		return
+	}
+
+	ctx.Response.Header.Add("Content-Type", "application/json")
+	ctx.Response.SetStatusCode(http.StatusOK)
+	ctx.Response.SetBody(data)
+}
+
 func (server *server) apiGetDiagnosticsHandler(ctx *fasthttp.RequestCtx) {
 	report, err := diagnostics.GenerateReport()
 	if err != nil {
@@ -585,6 +598,8 @@ func (server *server) Start() error {
 		// Interpretations
 		api.GET("/pods/{pod}/interpretations", apiGetInterpretationsHandler)
 		api.POST("/pods/{pod}/interpretations", apiPostInterpretationsHandler)
+
+		api.GET("/algorithms", server.apiGetAlgorithmsHandler)
 
 		api.GET("/diagnostics", server.apiGetDiagnosticsHandler)
 	}
